organise test helper code
